chore: refactor projectservice to use DTO's
