--- ProjectAlchemy.Core/Dtos/ProjectView.cs
@@ -1,22 +1,0 @@
-using ProjectAlchemy.Core.Domain;
-
-namespace ProjectAlchemy.Core.Dtos;
-
-public class ProjectView
-{
-    public required string Id { get; init; }
-    public required string Name { get; init; }
-    public required IEnumerable<LaneView> Lanes { get; init; }
-    public required IEnumerable<PartialIssue> Issues { get; init; }
-    
-    public static ProjectView FromProject(Project p)
-    {
-        return new ProjectView()
-        {
-            Id = p.Id,
-            Name = p.Name,
-            Lanes = p.Lanes.Select(LaneView.FromLane),
-            Issues = p.Issues.Select(PartialIssue.FromIssue),
-        };
-    }
-}
