+++ ProjectAlchemy.Core/Interfaces/IMemberRepository.cs
@@ -1,5 +1,6 @@
 using ProjectAlchemy.Core.Domain;
 using ProjectAlchemy.Core.Dtos;
+using ProjectAlchemy.Core.Dtos.Project;
 
 namespace ProjectAlchemy.Core.Interfaces;
