+++ ProjectAlchemy.Core/Interfaces/IProjectRepository.cs
@@ -1,4 +1,6 @@
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Dtos;
+using ProjectAlchemy.Core.Dtos.Project;
 
 namespace ProjectAlchemy.Core.Interfaces;
