+++ ProjectAlchemy.Core/Services/ProjectService.cs
@@ -1,4 +1,6 @@
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Dtos;
+using ProjectAlchemy.Core.Dtos.Project;
 using ProjectAlchemy.Core.Enums;
 using ProjectAlchemy.Core.Exceptions;
 using ProjectAlchemy.Core.Interfaces;
@@ -7,6 +9,7 @@ namespace ProjectAlchemy.Core.Services;
 
 public class ProjectService(IProjectRepository repository, AuthorizationService authService)
 {
+    public const int MaxNameLength = 30;
     private readonly IReadOnlyList<Lane> _defaultLanes = [new("To do"), new("In progress"), new("Done")];
 
     public async Task<Project> Get(string projectId, string userid)
@@ -24,7 +27,14 @@ public class ProjectService(IProjectRepository repository, AuthorizationService 
     public async Task<Project> Create(string projectName, string userId)
     {
         var creator = new Member(userId, MemberType.Owner);
-        var project = new Project(projectName, [], [creator], _defaultLanes.ToList());
+        var project = new Project
+        {
+            Id = Guid.NewGuid().ToString(),
+            Name = projectName,
+            Issues = [],
+            Lanes = _defaultLanes,
+            Members = [creator]
+        };
         return await repository.Create(project);
     }
 }
