+++ ProjectAlchemy.Core/Services/UserService.cs
@@ -1,5 +1,6 @@
 using ProjectAlchemy.Core.Domain;
 using ProjectAlchemy.Core.Dtos;
+using ProjectAlchemy.Core.Dtos.Project;
 using ProjectAlchemy.Core.Interfaces;
 
 namespace ProjectAlchemy.Core.Services;
